+++ ts/core/connector/bezier-connector.ts
@@ -2,14 +2,13 @@ import {AbstractBezierConnector, AbstractBezierOptions} from "./abstract-bezier-
 import {PaintGeometry, ConnectorComputeParams} from "./abstract-connector"
 import { JsPlumbInstance } from "../core"
 import {BezierSegment} from "./bezier-segment"
-import {Connectors} from "./connectors"
 import {Connection} from "./connection-impl"
 import {AnchorPlacement} from "../router/router"
 
-export class Bezier extends AbstractBezierConnector {
+export class BezierConnector extends AbstractBezierConnector {
 
     static type = "Bezier"
-    type = Bezier.type
+    type = BezierConnector.type
 
     majorAnchor:number
     minorAnchor:number